QC-cc

+++ src/main/java/com/qa/cinema/persistence/SeatType.java
@@ -2,6 +2,7 @@ package com.qa.cinema.persistence;
 
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 import javax.persistence.*;
@@ -26,7 +27,7 @@ public class SeatType implements Serializable {
 	private String name;
 
 	@Column(name="type_price")
-	private double price;
+	private BigDecimal price;
 
 	//bi-directional many-to-one association to Seat
 	@OneToMany(mappedBy="seatType", cascade={CascadeType.ALL})
@@ -35,7 +36,7 @@ public class SeatType implements Serializable {
 	public SeatType() {
 	}
 	
-	public SeatType(Long typeId, String name, double price) {
+	public SeatType(Long typeId, String name, BigDecimal price) {
 		this.typeId = typeId;
 		this.name = name;
 		this.price = price;
@@ -57,11 +58,11 @@ public class SeatType implements Serializable {
 		this.name = name;
 	}
 
-	public double getPrice() {
+	public BigDecimal getPrice() {
 		return this.price;
 	}
 
-	public void setPrice(double price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
